refactor: help, 알림 설정 리팩토링
