+++ src/components/my/data/faqs.ts
@@ -1,0 +1,51 @@
+import type { FAQ } from "@/components/my/types/help";
+
+export const FAQS: FAQ[] = [
+  {
+    category: "일반",
+    question: "PopUp!은 무엇인가요?",
+    answer:
+      "PopUp!은 전국의 팝업스토어 정보를 한눈에 확인하고, 나만의 팝업 여행을 계획할 수 있는 서비스입니다. 실시간으로 업데이트되는 팝업 정보를 확인하고, 관심 있는 팝업을 저장해보세요!",
+  },
+  {
+    category: "기능",
+    question: "팝업을 저장하려면 어떻게 하나요?",
+    answer:
+      "팝업 카드의 하트 아이콘을 탭하면 저장됩니다. 저장된 팝업은 마이페이지에서 확인할 수 있어요.",
+  },
+  {
+    category: "기능",
+    question: "알림은 어떻게 설정하나요?",
+    answer:
+      "마이페이지 > 알림 설정에서 원하는 알림 종류를 선택할 수 있습니다. 새로운 팝업, 근처 팝업, 일정 알림 등 다양한 알림을 제공해요.",
+  },
+  {
+    category: "캡슐",
+    question: "캡슐 뽑기는 무엇인가요?",
+    answer:
+      "오늘의 팝업 운세를 확인할 수 있는 재미있는 기능이에요! 하루에 한 번 캡슐을 뽑아 추천 팝업과 귀여운 마스코트를 만나보세요.",
+  },
+  {
+    category: "제보",
+    question: "새로운 팝업을 제보하고 싶어요",
+    answer:
+      "마이페이지 > 팝업 제보하기에서 새로운 팝업 정보를 공유해주세요. 여러분의 제보가 PopUp! 커뮤니티를 더욱 풍성하게 만듭니다!",
+  },
+  {
+    category: "계정",
+    question: "프로필 정보를 변경하려면?",
+    answer: "마이페이지 > 프로필 편집에서 이름, 이메일, 위치 등의 정보를 수정할 수 있습니다.",
+  },
+  {
+    category: "일반",
+    question: "지도에서 팝업을 찾으려면?",
+    answer:
+      "지도 탭에서 현재 위치 주변의 팝업을 확인할 수 있어요. 색상별로 카테고리가 구분되어 있어 원하는 팝업을 쉽게 찾을 수 있습니다.",
+  },
+  {
+    category: "기능",
+    question: "여행 로드맵은 어떻게 사용하나요?",
+    answer:
+      "로드맵 탭에서 방문 예정인 팝업들의 일정을 타임라인으로 확인할 수 있어요. 효율적인 팝업 투어를 계획해보세요!",
+  },
+];
